requirements.txt
asgiref==3.5.0
black==22.1.0
click==8.0.4
colorama==0.4.4
Django==4.0.3
django-taggit==2.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
tzdata==2021.5
